requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
azure-ai-agents==1.0.2
azure-ai-evaluation==1.11.0
azure-ai-ml==1.28.1
azure-ai-projects==1.0.0
azure-common==1.1.28
azure-core==1.35.1
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.0
azure-mgmt-core==1.6.0
azure-monitor-opentelemetry==1.6.13
azure-monitor-opentelemetry-exporter==1.0.0b41
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
azure-storage-file-share==12.22.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
distro==1.9.0
fixedint==0.1.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
isodate==0.7.2
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markupsafe==3.0.2
marshmallow==3.26.1
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.4
nltk==3.9.1
numpy==2.3.3
oauthlib==3.3.1
openai==1.108.0
opentelemetry-api==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-dbapi==0.57b0
opentelemetry-instrumentation-django==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-flask==0.57b0
opentelemetry-instrumentation-openai-v2==2.1b0
opentelemetry-instrumentation-psycopg2==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-urllib==0.57b0
opentelemetry-instrumentation-urllib3==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
pandas==2.3.2
propcache==0.3.2
psutil==7.0.0
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydash==8.0.5
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.26.0
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.12
six==1.17.0
sniffio==1.3.1
strictyaml==1.7.3
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
